pyproject.toml
[tool.poetry]
name = "label"
version = "0.1.0"
description = ""
authors = ["Massimiliano Ravelli <massimiliano.ravelli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^4.2.7"
factory-boy = "^3.3.0"
django-colorfield = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
